refactor: extract smithy-send-patch into shared file and removed contract test leftovers
