+++ aws-distro-opentelemetry-node-autoinstrumentation/src/patches/instrumentation-patch.ts
@@ -11,11 +11,9 @@ import {
   trace,
   Span,
   Tracer,
-  defaultTextMapSetter,
 } from '@opentelemetry/api';
 import { Instrumentation } from '@opentelemetry/instrumentation';
 import {
-  AwsInstrumentation,
   AwsSdkInstrumentationConfig,
   NormalizedRequest,
   NormalizedResponse,
@@ -32,9 +30,7 @@ import {
 } from './aws/services/bedrock';
 import { BedrockAgentCoreServiceExtension } from './aws/services/bedrock-agentcore';
 import type { AwsLambdaInstrumentation } from '@opentelemetry/instrumentation-aws-lambda';
-import type { Command as AwsV3Command } from '@aws-sdk/types';
 import { LoggerProvider } from '@opentelemetry/api-logs';
-import { suppressTracing } from '@opentelemetry/core';
 
 export const traceContextEnvironmentKey = '_X_AMZN_TRACE_ID';
 export const AWSXRAY_TRACE_ID_HEADER_CAPITALIZED = 'X-Amzn-Trace-Id';
@@ -456,105 +452,9 @@ function patchAwsLambdaInstrumentation(instrumentation: Instrumentation): void {
   }
 }
 
-// Override the upstream private _getV3SmithyClientSendPatch method to add middlewares to inject X-Ray Trace Context into HTTP Headers and to extract account access key id and region for cross-account support
-// https://github.com/open-telemetry/opentelemetry-js-contrib/blob/instrumentation-aws-sdk-v0.48.0/plugins/node/opentelemetry-instrumentation-aws-sdk/src/aws-sdk.ts#L373-L384
-const V3_CLIENT_CONFIG_KEY = Symbol('opentelemetry.instrumentation.aws-sdk.client.config');
-type V3PluginCommand = AwsV3Command<any, any, any, any, any> & {
-  [V3_CLIENT_CONFIG_KEY]?: any;
-};
-// Symbol to prevent infinite recursion during credential capture
-// When we extract credentials, the AWS SDK may need to make additional AWS API calls
-// (e.g., sts:AssumeRoleWithWebIdentity) which go through the same instrumented 'send' method.
-// Without this flag, each credential request would trigger another credential extraction attempt,
-// creating an infinite loop of nested AWS SDK calls.
-export const SKIP_CREDENTIAL_CAPTURE_KEY = Symbol('skip-credential-capture');
+import { applySmithySendPatch, SKIP_CREDENTIAL_CAPTURE_KEY } from './smithy-send-patch';
+export { SKIP_CREDENTIAL_CAPTURE_KEY };
+
 function patchAwsSdkInstrumentation(instrumentation: Instrumentation): void {
-  if (instrumentation) {
-    (instrumentation as AwsInstrumentation)['_getV3SmithyClientSendPatch'] = function (
-      moduleVersion: string | undefined,
-      original: (...args: unknown[]) => Promise<any>
-    ) {
-      const self = this;
-      return function send(this: any, command: V3PluginCommand, ...args: unknown[]): Promise<any> {
-        // Only add middleware once per client instance to reduce overhead
-        // AWS SDK clients may call 'send' multiple times, but we only need to patch once
-        // Even with override=true, adding middleware still causes overhead as it replaces existing stack entries
-        if (!this.__adotMiddlewarePatched) {
-          this.middlewareStack?.add(
-            (next: any, context: any) => async (middlewareArgs: any) => {
-              propagation.inject(otelContext.active(), middlewareArgs.request.headers, defaultTextMapSetter);
-              // Need to set capitalized version of the trace id to ensure that the Recursion Detection Middleware
-              // of aws-sdk-js-v3 will detect the propagated X-Ray Context
-              // See: https://github.com/aws/aws-sdk-js-v3/blob/v3.768.0/packages/middleware-recursion-detection/src/index.ts#L13
-              const xrayTraceId = middlewareArgs.request.headers[AWSXRAY_TRACE_ID_HEADER];
-
-              if (xrayTraceId) {
-                middlewareArgs.request.headers[AWSXRAY_TRACE_ID_HEADER_CAPITALIZED] = xrayTraceId;
-                delete middlewareArgs.request.headers[AWSXRAY_TRACE_ID_HEADER];
-              }
-              const result = await next(middlewareArgs);
-              return result;
-            },
-            {
-              step: 'build',
-              name: '_adotInjectXrayContextMiddleware',
-              override: true,
-            }
-          );
-
-          this.middlewareStack?.add(
-            (next: any, context: any) => async (middlewareArgs: any) => {
-              const activeContext = otelContext.active();
-              // Skip credential extraction if this is a nested call from another credential extraction
-              // This prevents infinite recursion when credential providers make AWS API calls
-              if (activeContext.getValue(SKIP_CREDENTIAL_CAPTURE_KEY)) {
-                return await next(middlewareArgs);
-              }
-              const span = trace.getSpan(activeContext);
-
-              if (span) {
-                // suppressTracing prevents span generation for internal credential extraction calls
-                // which are implementation details and not relevant to the application's telemetry
-                const suppressedContext = suppressTracing(activeContext).setValue(SKIP_CREDENTIAL_CAPTURE_KEY, true);
-                // Skip credential extraction if the context is not injectable
-                if (suppressedContext.getValue(SKIP_CREDENTIAL_CAPTURE_KEY)) {
-                  await otelContext.with(suppressedContext, async () => {
-                    try {
-                      const credsProvider = this.config.credentials;
-                      if (credsProvider instanceof Function) {
-                        const credentials = await credsProvider();
-                        if (credentials?.accessKeyId) {
-                          span.setAttribute(AWS_ATTRIBUTE_KEYS.AWS_AUTH_ACCOUNT_ACCESS_KEY, credentials.accessKeyId);
-                        }
-                      }
-                      if (this.config.region instanceof Function) {
-                        const region = await this.config.region();
-                        if (region) {
-                          span.setAttribute(AWS_ATTRIBUTE_KEYS.AWS_AUTH_REGION, region);
-                        }
-                      }
-                    } catch (err) {
-                      diag.debug('Failed to get auth account access key and region:', err);
-                    }
-                  });
-                }
-              }
-
-              return await next(middlewareArgs);
-            },
-            {
-              step: 'build',
-              name: '_adotExtractSignerCredentials',
-              override: true,
-            }
-          );
-          this.__adotMiddlewarePatched = true;
-        }
-
-        command[V3_CLIENT_CONFIG_KEY] = this.config;
-        self['patchV3MiddlewareStack'](moduleVersion, this.middlewareStack);
-        return original.apply(this, [command, ...args]);
-      };
-    };
-  }
+  applySmithySendPatch(instrumentation);
 }
